Make sure that all enums have a string representation
This was not checked before. We cannot simply put the string
representation or the key to the enum keys themselves, since this would
invalidate our ResourceBundle tests. Hence we need to have a test to link
the two aspects: Enum keys and String representation. The representation
is now moved to the enums themselves